requirements.txt
atproto==0.0.31
mysql-connector-python==8.2.0
python-dotenv==1.0.0
schedule==1.2.1
python-logging-handler==1.0.8
pytest==7.4.3
